requirements.txt
click==7.1.1
cssmin==0.2.0
Flask==1.1.1
Flask-Assets==2.0
itsdangerous==1.1.0
Jinja2==2.11.1
jsmin==2.2.2
lesscpy==0.14.0
MarkupSafe==1.1.1
ply==3.11
python-dotenv==0.12.0
six==1.14.0
webassets==2.0
Werkzeug==1.0.0
PyMongo==3.3.0
flask_pymongo==2.3.0
requests==2.23.0
boto3>=1.12.38
botocore<=1.16.0
sshtunnel>=0.1.5
colorama>=0.4.3
termcolor>=1.1.0
flask-cors==3.0.8
fabric2<=2.5.0
elasticsearch<=7.8.1
requests-aws4auth<=1.0
aws-requests-auth==0.4.3
retry==0.9.2
lxml<=4.5.2
